+++ punchpipe/control/cache_layer/manager.py
@@ -2,6 +2,8 @@ import contextlib
 from glob import glob
 from multiprocessing.shared_memory import SharedMemory
 
+from prefect import get_run_logger
+from prefect.exceptions import MissingContextError
 from prefect.variables import Variable
 
 CACHE_KEY_PREFIX = "punchpipe-cache-"
@@ -47,6 +49,10 @@ def try_write_to_key(key, data):
         shm.buf[0] = 0
         shm.buf[1:len(data)+1] = data
         shm.buf[0] = 1
+        try:
+            get_run_logger().info(f"Saved to cache key {key}")
+        except MissingContextError:
+            pass  # we're not in a flow so we don't log
     except FileExistsError:
         pass
     finally: